requirements.txt
# requirements.txt
mss
keyboard
pygame
pytesseract
pillow
edge-tts
pywin32